[MH-260] Make links on upload page go to their faq entry.
I also provided a new tab opener for the general requirements since we
actually have faq entries we can point these links to. This will of
course change when we want to have hover links - but would I think be
more helpful thank a non-functioning link during the demos at the
conference this week.

Also simplified the anchor tag on() interception that jquery does (just
don't register listeners for those tags that have a data-no-ajax
property)