[MH-281] Adds flake8 linting.
I've gone ahead and included fabric with this, as we use that in
automation for other projects - and created a command to lint, check for
un-added migrations, and run tests (note that this is v2 of fabric, the
latest and greatest).

Warnings were mostly whitespace, and argument order. I ignored the
operator order and cleaned up the remaining warnings.

Also included an editorconfig rc file with the defaults used at
Caktus (pretty much all editors have a plugin for this AFAIK).

The actual CI should change with this PR to make use of the new test.
The new steps to do CI should be:

```
./hsctl rebuild --db

docker exec -u hydro-service hydroshare fab test

./hsctl stop

docker-compose rm -f
```